завершает выделение представлений из модулей экранов
