+++ js/game/dicts.js
@@ -13,3 +13,26 @@ export const ANSWER_MODIFIER_MAP = {
   [ANSWER_TYPE.SLOW]: `slow`,
   [ANSWER_TYPE.WRONG]: `wrong`,
 };
+
+export const QUESTION_TYPE = {
+  'SINGLE': 1,
+  'DOUBLE': 2,
+  'TRIPLE': 3,
+};
+
+export const IMAGE_TYPE = {
+  'PHOTO': `photo`,
+  'PAINT': `paint`,
+};
+
+export const QUESTION_HAS_ANSWERS_MAP = {
+  [QUESTION_TYPE.SINGLE]: true,
+  [QUESTION_TYPE.DOUBLE]: true,
+  [QUESTION_TYPE.TRIPLE]: false,
+};
+
+export const QUESTION_MODIFIER_MAP = {
+  [QUESTION_TYPE.SINGLE]: `game__content--wide`,
+  [QUESTION_TYPE.DOUBLE]: ``,
+  [QUESTION_TYPE.TRIPLE]: `game__content--triple`,
+};
