+++ js/screens/game-3.js
@@ -1,30 +1,18 @@
-import createFragmentFromTemplate from '../dom-factory';
-import {addDelegatedEventListener} from '../utils';
 import {goToNextStep, saveAnswerResult, changeLivesCount} from '../game/game-logic';
-import createGameScreenMarkup from './game-screen';
 import {isAnswerCorrect} from '../game/is-answer-correct';
+import TripleImageView from '../views/triple-image-view';
 
 const createTripleScreenFragment = (state) => {
-  const HAS_ANSWERS = false;
-  const GAME_CONTENT_MODIFIER = `game__content--triple`;
+  const tripleImageView = new TripleImageView(state);
+  tripleImageView.onAnswer = (answer) => {
+    const correctAnswer = state.steps[state.currentStepIndex].options;
+    const answerResult = isAnswerCorrect(answer, correctAnswer);
+    let newState = saveAnswerResult(state, {correct: answerResult, timeElapsed: 15});
+    newState = changeLivesCount(newState, answerResult);
+    goToNextStep(newState);
+  };
 
-  const fragmentMarkup = createGameScreenMarkup(state, HAS_ANSWERS, GAME_CONTENT_MODIFIER);
-
-  const tripleScreenFragment = createFragmentFromTemplate(fragmentMarkup);
-  const gameContentForm = tripleScreenFragment.querySelector(`.game__content`);
-  addDelegatedEventListener(`click`, `.game__option`, (evt) => {
-    const checkedAnswerElement = evt.target;
-    if (checkedAnswerElement) {
-      const candidateAnswers = [JSON.parse(checkedAnswerElement.dataset[`answer`])];
-      const correctAnswers = state.steps[state.currentStepIndex].options;
-      const answerResult = isAnswerCorrect(candidateAnswers, correctAnswers);
-      let newState = saveAnswerResult(state, {correct: answerResult, timeElapsed: 15});
-      newState = changeLivesCount(newState, answerResult);
-      goToNextStep(newState);
-    }
-  }, gameContentForm);
-
-  return tripleScreenFragment;
+  return tripleImageView.element;
 };
 
 export default createTripleScreenFragment;
