+++ js/views/double-image-view.js
@@ -2,7 +2,7 @@ import {AbstractView} from "./abstract-view";
 import {addDelegatedEventListener} from "../utils";
 import createGameScreenMarkup from "../screens/game-screen";
 
-const SingleImageView = class extends AbstractView {
+const DoubleImageView = class extends AbstractView {
 
   constructor(state) {
     super();
@@ -17,15 +17,20 @@ const SingleImageView = class extends AbstractView {
     const gameElement = this.element.querySelector(`.game`);
 
     addDelegatedEventListener(`change`, `.game__content`, () => {
-      const checkedAnswerElement = gameElement.querySelector(`[name=question1]:checked`);
-      if (checkedAnswerElement) {
-        const answer = [JSON.parse(checkedAnswerElement.dataset[`answer`])];
+      const checkedQuestion1Element = document.querySelector(`[name=question1]:checked`);
+      const checkedQuestion2Element = document.querySelector(`[name=question2]:checked`);
+      if (checkedQuestion1Element && checkedQuestion2Element) {
+        const answer = [
+          JSON.parse(checkedQuestion1Element.dataset[`answer`]),
+          JSON.parse(checkedQuestion2Element.dataset[`answer`]),
+        ];
         this.onAnswer(answer);
       }
     }, gameElement);
   }
 
   onAnswer() {}
+
 };
 
-export default SingleImageView;
+export default DoubleImageView;
